% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-PLSRBin.R
\name{NPLSRBin}
\alias{NPLSRBin}
\title{Perform Non-linear Partial Least Squares Regression (NPLSR) for
Binary Responses}
\usage{
NPLSRBin(
  Y,
  X,
  I,
  J,
  K,
  L,
  inames = NULL,
  xnames = NULL,
  ynames = NULL,
  S = 2,
  tolerance = 5e-05,
  maxiter = 100,
  penalization = 0.1,
  OptimMethod = "CG"
)
}
\arguments{
\item{Y}{Response matrix (binary) of dimension I x L.}

\item{X}{Matrix of predictors (explanatory variables) of dimension
I x J*K.}

\item{I}{Number of observations.}

\item{J}{Number of columns in the first dimension of the original
matrix of predictors.}

\item{K}{Number of columns in the second dimension of the original
matrix of predictors.}

\item{L}{Number of response variables.}

\item{inames}{Names for the observations (optional, default is NULL).}

\item{xnames}{Names for the predictors (optional, default is NULL).}

\item{ynames}{Names for the response variables (optional, default is NULL).}

\item{S}{Number of latent variables (default is 2).}

\item{tolerance}{Tolerance value for convergence (default is 0.00005).}

\item{maxiter}{Maximum number of iterations (default is 100).}

\item{penalization}{Penalization parameter (default is 0.1).}

\item{OptimMethod}{Optimization method (default is "CG").}
}
\value{
An object of class 'PLSRBin' containing various components of the
NPLSR model including scores, loadings, coefficients, predictions, and percent correct classifications.
}
\description{
This function conducts Non-linear Partial Least Squares Regression
(NPLSR) for binary response variables.
}
\examples{
X<- as.matrix(X3way)
pls <-NPLSRBin(Y,X,8,7,6,9)
plsbip=Biplot.PLSRBIN(pls)
plot(plsbip)

}
