% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-PLSRBinFit.R
\name{NPLSRBinFit}
\alias{NPLSRBinFit}
\title{Fit Non-linear Partial Least Squares Regression (NPLSR) Model for Binary
Response}
\usage{
NPLSRBinFit(
  Y,
  X,
  I,
  J,
  K,
  L,
  S = 2,
  tolerance = 5e-06,
  maxiter = 100,
  penalization = 0.1,
  OptimMethod = "CG"
)
}
\arguments{
\item{Y}{Response matrix (binary) of dimension I x L.}

\item{X}{Matrix of predictors (explanatory variables) of dimension
I x J*K.}

\item{I}{Number of observations.}

\item{J}{Number of columns in the first dimension of the original
matrix of predictors.}

\item{K}{Number of columns in the second dimension of the original
matrix of predictors.}

\item{L}{Number of response variables.}

\item{S}{Number of latent variables (default is 2).}

\item{tolerance}{Tolerance value for convergence (default is 0.000005).}

\item{maxiter}{Maximum number of iterations (default is 100).}

\item{penalization}{Penalization parameter (default is 0.1).}

\item{OptimMethod}{Optimization method (default is "CG").}
}
\value{
A list containing various components of the NPLSR model including
latent variables, loadings, predictions, and percent correct classifications.
}
\description{
This function fits a Non-linear Partial Least Squares Regression
(NPLSR) model for binary response variables.
}
\examples{
X<- as.matrix(X3way)
pls <-NPLSRBinFit(Y,X,8,7,6,9)
}
